Fix: Move is_sync_point_signature_verified to common/bls
